requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astrapy==2.2.1
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.8.4
decorator==5.2.1
deprecation==2.1.0
dill==0.4.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.1
filelock==3.25.2
Flask==3.1.3
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.4.0
groq==0.37.1
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.2
hyperframe==6.1.0
idna==3.11
ipython==9.12.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.3.25
langchain-astradb==1.0.0
langchain-community==0.3.24
langchain-core==0.3.84
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.3.45
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
numpy==2.4.4
orjson==3.11.8
packaging==25.0
pandas==3.0.2
parso==0.8.6
prompt_toolkit==3.0.52
propcache==0.4.1
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
stack-data==0.6.3
tenacity==9.1.4
tokenizers==0.22.2
toml==0.10.2
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid6==2025.0.1
uuid_utils==0.14.1
wcwidth==0.6.0
Werkzeug==3.1.8
xxhash==3.6.0
yarl==1.23.0
zstandard==0.23.0
streamlit